Create deleteAccount and useDeletionServices
