+++ client/src/features/login/api/__mocks__/account-api.ts
@@ -1,7 +1,13 @@
-export const mockFetchUnderstudies = jest.fn();
+export const mockLogin = jest.fn();
+export const mockRegister = jest.fn();
+export const mockLogout = jest.fn();
+export const mockDeleteAccount = jest.fn();
 
 const mock = () => ({
-    fetchUnderstudies: mockFetchUnderstudies
+    login: mockLogin,
+    register: mockRegister,
+    logout: mockLogout,
+    deleteAccount: mockDeleteAccount,
 });
 
 export default mock;
