--- client/src/features/login/api/__mocks__/auth-api.ts
@@ -1,9 +1,0 @@
-export const mockLogin = jest.fn();
-export const mockRegister = jest.fn();
-export const mockLogout = jest.fn();
-
-export const AuthApi = () => ({
-    login: mockLogin,
-    register: mockRegister,
-    logout: mockLogout
-});
