+++ client/src/features/login/api/__mocks__/understudy-api.ts
@@ -1,0 +1,7 @@
+export const mockFetchUnderstudies = jest.fn();
+
+const mock = () => ({
+    fetchUnderstudies: mockFetchUnderstudies
+});
+
+export default mock;
