+++ client/src/features/login/components/Login.tsx
@@ -3,7 +3,7 @@ import { Link, useNavigate, useLocation } from "react-router-dom";
 import { useHotKeys } from "../../../hooks/useHotKeys";
 import { IconButton, InputAdornment, TextField } from "@mui/material";
 import { Visibility, VisibilityOff } from "@mui/icons-material";
-import { useLoginService } from "../hooks/useAuth";
+import { useLoginService } from "../hooks/useAccountServices";
 
 const LoginScreen = () => {
   const [email, setEmail] = useState("");
