+++ client/src/hooks/__mocks__/useApiConnection.ts
@@ -1,11 +1,11 @@
 export const mockGet = jest.fn();
 export const mockPost = jest.fn();
 export const mockPut = jest.fn();
-export const mockDelete = jest.fn();
+export const mockDel = jest.fn();
 
 export const useApiConnection = () => ({
     get: mockGet,
     post: mockPost,
     put: mockPut,
-    delete: mockDelete
+    del: mockDel
 });
